+++ android/src/main/java/com/figmashadow/FigmaShadowViewPackage.kt
@@ -3,7 +3,6 @@ package com.figmashadow
 import com.facebook.react.BaseReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.module.model.ReactModuleInfo
 import com.facebook.react.module.model.ReactModuleInfoProvider
 import com.facebook.react.uimanager.ViewManager
 
@@ -15,16 +14,5 @@ class FigmaShadowViewPackage : BaseReactPackage() {
     reactContext: ReactApplicationContext,
   ): List<ViewManager<*, *>> = listOf(FigmaShadowViewManager())
 
-  override fun getReactModuleInfoProvider(): ReactModuleInfoProvider = ReactModuleInfoProvider {
-    mapOf(
-      FigmaShadowViewManager.NAME to ReactModuleInfo(
-        FigmaShadowViewManager.NAME,
-        FigmaShadowViewManager.NAME,
-        false, // canOverrideExistingModule
-        false, // needsEagerInit
-        false, // isCxxModule
-        true, // isTurboModule
-      ),
-    )
-  }
+  override fun getReactModuleInfoProvider() = ReactModuleInfoProvider { emptyMap() }
 }